==> 00-download-dados-setores-2000.ipynb <==
import numpy as np
from ftplib import FTP
import urllib.request
import zipfile
import os.path
import glob
import shutil

URL_DOWNLOAD = 'https://ftp.ibge.gov.br/Censos/Censo_Demografico_2000/Dados_do_Universo/Agregado_por_Setores_Censitarios/'

ftp = FTP('ftp.ibge.gov.br')
ftp.login()
ftp.cwd('Censos/Censo_Demografico_2000/Dados_do_Universo/Agregado_por_Setores_Censitarios/')
files = ftp.nlst()
ftp.quit()


files

UFS = [
    'AC',
    'AL',
    'AM',
    'AP',
    'BA',
    'CE',
    'DF',
    'ES',
    'GO',
    'MA',
    'MG',
    'MS',
    'MT',
    'PA',
    'PB',
    'PE',
    'PI',
    'PR',
    'RJ',
    'RN',
    'RO',
    'RR',
    'RS',
    'SC',
    'SE',
    'SP_Exclusive_RM',
    'SP_RM',
    'TO'
]

np.array(files)[np.char.startswith(files, 'Agregado_de_setores_2000_SP_Exclusive_RM')][0]

for uf in UFS:
    starts_with = np.char.startswith(files, 'Agregado_de_setores_2000_' + uf)
    file = np.array(files)[starts_with][0]
    url = f'{URL_DOWNLOAD}/{file}'
    file_target = f'data/censo_2000/{uf}.zip'
    if not os.path.isfile(file_target):
        urllib.request.urlretrieve(url, f'data/censo_2000/{uf}.zip')
    print(file) 

for uf in UFS:
    print(f'Descompactando {uf}')
    folder = f'data/censo_2000/csvs'
    check_folder = os.path.isdir(folder)
    if not check_folder:
        os.makedirs(folder)
    os.makedirs('temp')
    with zipfile.ZipFile(f'data/censo_2000/{uf}.zip', 'r') as zip_ref:
        zip_ref.extractall('temp/')
        csv_files = glob.glob(f'temp/**/*.XLS', recursive=True)
        for f in csv_files:
            shutil.copyfile(f, f'{folder}/{f.split("/")[-1]}')
    shutil.rmtree('temp')
    # break

## Arrumando nomes de arquivos SP2
#[os.rename(f, '_'.join(f.split('_')[:2]) + '_SP2.csv') for f in glob.glob('data/censo_2010/csvs/*sp2.csv')]

#[os.rename(f, f.replace('SP', 'SP2')) for f in glob.glob('data/censo_2010/csvs/*SP.csv')]

#[os.rename(f, f.replace('responsavel', 'Responsavel')) for f in glob.glob('data/censo_2010/csvs/responsavel*.csv')]
