==> 00-extraindo-area-dos-setores-2000.ipynb <==
import geopandas as gpd
import pandas as pd
import glob

# Downloado dos dados da malha censitária de 2010
#

# Download das geometrias da malha censitária de 2000
# https://github.com/bplmp/shapefiles-setores-censitarios-censo-2000/tree/master

setores = glob.glob("data/censo_2000/*_setores_censitarios.zip")

setores

areas = []
gdfs = []

for setor in setores:
    gdf_uf = gpd.read_file(setor)
    gdf_uf.to_crs(epsg="5880", inplace=True)
    gdf_uf.loc[:, 'area_do_setor'] = gdf_uf.area
    gdfs.append(gdf_uf)
    areas.append(gdf_uf.loc[:, ['CD_GEOCODI', 'area_do_setor']])
    # break

gdf_setores_br = pd.concat(gdfs)

gdf_setores_br.shape

gdf_setores_br

gdf_setores_br.to_file('data/censo_2000/BR_Setores_2000.gpkg', driver='GPKG')

pd.concat(areas).set_index('CD_GEOCODI').to_csv('resultados/area_dos_setores_2000.zip')

pd.read_csv('resultados/area_dos_setores_2000.zip').dtypes

ufs = glob.glob("data/censo_2010/*_municipios.zip")

len(ufs)

# # Trabalhando os distritos
# # ufs = []
# gdfs = []

# for uf in ufs:
#     gdf_uf = gpd.read_file(uf, encoding='ISO-8859-15')
#     gdf_uf.to_crs(epsg="4674", inplace=True)
#     # gdf_uf.loc[:, 'area_do_setor'] = gdf_uf.area
#     gdfs.append(gdf_uf)
#     # break

# gdf_distritos = pd.concat(gdfs)

# gdf_distritos.to_file('data/censo_2010/municipios.gpkg', driver='GPKG')
