==> 01-preparando-os-dados-2010.ipynb <==
import pandas as pd
import geopandas as gpd
# from tabula import read_pdf

# df_informacoes = read_pdf('data/censo_2010/BASE DE INFORMAåES POR SETOR CENSITµRIO Censo 2010 - Universo.pdf', stream=True, pages=46, lattice=True)

# df_informacoes[0]

## Malha de setores do BR
# https://www.ibge.gov.br/geociencias/downloads-geociencias.html?caminho=organizacao_do_territorio/malhas_territoriais/malhas_de_setores_censitarios__divisoes_intramunicipais/2021/Malha_de_setores_(shp)_Brasil

## Download das tabelas e documentacao
# https://www.ibge.gov.br/estatisticas/sociais/populacao/9662-censo-demografico-2010.html?edicao=10410

# Selecao das variáveis

## Habitação einfraestrutura

* Porcentagem de domicílios do tipo cômodo
* Porcentagem de domicílios sem coleta de lixo
* Porcentagem de domicílios sem ligação à rede de abastecimento de água
* Porcentagem de domicílios sem banheiros ou sanitários
* Porcentagem de domicílios sem ligação à rede de esgoto ou fossa séptica
* Porcentagem de domicílios – outra forma de posse da moradia
* Número de banheiros por habitante

## Renda e escolaridade do responsável pelo domicílio

* Porcentagem de responsáveis por domicílio não alfabetizados
* Porcentagem de responsáveis por domicílio com menos de 30 anos não alfabetizados
* Porcentagem de responsáveis por domicílio com renda de até 3 salários mínimos
* Porcentagem de responsáveis por domicílio com menos de 8 anos de estudo
* Anos médios de estudo do responsável pelo domicílio
* Renda média do responsável pelo domicílio

## Aspectos demográficos

* Número de domicílios particulares permanentes no setor censitário
* Número de domicílios improvisados no setor censitário
* Número de pessoas residentes no setor censitário
* Porcentagem de responsáveis por domicílios com menos de 30 anos
* Número médio de pessoas por domicílio



# gdf_br = gpd.read_file('data/censo_2010/BR_Setores_2021.zip')

# gdf_br.shape

# gdf_br.NM_SIT.value_counts()

# gdf_br.columns

UFS = [
    'AC',
    'AL',
    'AM',
    'AP',
    'BA',
    'CE',
    'DF',
    'ES',
    'GO',
    'MA',
    'MG',
    'MS',
    'MT',
    'PA',
    'PB',
    'PE',
    'PI',
    'PR',
    'RJ',
    'RN',
    'RO',
    'RR',
    'RS',
    'SC',
    'SE',
    'SP1',
    'SP2',
    'TO'
]

## Hash para montar a base de dados
# {'file1': ['var1', 'var2', ..., 'var n'], ...,'file n': ['var1', 'var2', ..., 'var n']}

model_vars = {
    'Basico':{
        # 'Cod_setor': 'cod_setor',
        'Cod_Grandes Regiões': 'cod_grande_regiao',
        # 'Nome_Grande_Regiao': 'Nome GRande Região',
        'Cod_UF': 'uf',
        'Cod_meso': 'meso_regiao',
        'Cod_RM': 'rm',
        'Nome_da_RM': 'nome_rm',
        # 'Cod_Grandes Regiões': ''
        'Cod_municipio': 'cod_municipio',
        'Nome_do_municipio': 'nome_municipio',
        'Situacao_setor': 'situacao_setor',
        'Tipo_setor': 'tipo_do_setor',
        'V001': 'numero_de_domicilios',
        'V002': 'numero_moradores'
    },
    'Domicilio01': {
        'V003': 'domicilios_tipo_casa',
        'V002': 'domicilios_particulares_permanentes',
        'V035': 'domicilios_com_lixo_coletado',
        'V106': 'domicilios_tipo_casa_com_abastecimento_de_agua',
        'V023': 'domicilios_particulares_permanentes_sem_banheiro_ou_sanitario',
        'V017': 'domicilios_particulares_permanentes_com_banheiro_ligado_a_esgoto',
        'V018': 'domicilios_particulares_permanentes_com_banheiro_ligado_a_fossa_septica',
        'V025': 'domicilios_particulares_permanentes_com_1_banheiro',
        'V026': 'domicilios_particulares_permanentes_com_2_banheiros',
        'V027': 'domicilios_particulares_permanentes_com_3_banheiros',
        'V028': 'domicilios_particulares_permanentes_com_4_banheiros',
        'V029': 'domicilios_particulares_permanentes_com_5_banheiros',
        'V030': 'domicilios_particulares_permanentes_com_6_banheiros',
        'V031': 'domicilios_particulares_permanentes_com_7_banheiros',
        'V032': 'domicilios_particulares_permanentes_com_8_banheiros',
        'V033': 'domicilios_particulares_permanentes_com_mais_de_9_banheiros',
        'V010': 'domicilios_particulares_permanentes_cedidos_de_outra_forma',
        'V011': 'domicilios_particulares_permanentes_em_outra_condicao_de_ocupacao',
    },
    'Responsavel02': {
        'V001': 'pessoas_responsaveis',
        'V002': 'pessoas_responsaveis_com_10_anos',
        'V003': 'pessoas_responsaveis_com_11_anos',
        'V004': 'pessoas_responsaveis_com_12_anos',
        'V005': 'pessoas_responsaveis_com_13_anos',
        'V006': 'pessoas_responsaveis_com_14_anos',
        'V007': 'pessoas_responsaveis_com_15_anos',
        'V008': 'pessoas_responsaveis_com_16_anos',
        'V009': 'pessoas_responsaveis_com_17_anos',
        'V010': 'pessoas_responsaveis_com_18_anos',
        'V011': 'pessoas_responsaveis_com_19_anos',
        'V012': 'pessoas_responsaveis_com_20_anos',
        'V013': 'pessoas_responsaveis_com_21_anos',
        'V014': 'pessoas_responsaveis_com_22_anos',
        'V015': 'pessoas_responsaveis_com_23_anos',
        'V016': 'pessoas_responsaveis_com_24_anos',
        'V017': 'pessoas_responsaveis_com_25_anos',
        'V018': 'pessoas_responsaveis_com_26_anos',
        'V019': 'pessoas_responsaveis_com_27_anos',
        'V020': 'pessoas_responsaveis_com_28_anos',
        'V021': 'pessoas_responsaveis_com_29_anos',
        'V093': 'pessoas_responsaveis_alfabetizadas',
        'V094': 'pessoas_responsaveis_alfabetizadas_de_10_a_14_anos',
        'V095': 'pessoas_responsaveis_alfabetizadas_de_15_a_19_anos',
        'V096': 'pessoas_responsaveis_alfabetizadas_de_20_a_24_anos',
        'V097': 'pessoas_responsaveis_alfabetizadas_de_25_a_29_anos'
    },
    'DomicilioRenda':{
        'V001': 'total_domicilios_improvisados',
        'V002': 'total_rendimento_dos_domicilios',
    },
    'ResponsavelRenda':{
        'V001': 'responsaveis_com_renda_ate_meio_salario_minimo',
        'V002': 'responsaveis_com_renda_de_meio_a_1_salario_minimo',
        'V003': 'responsaveis_com_renda_de_1_a_2_salarios_minimos',
        'V004': 'responsaveis_com_renda_de_2_a_3_salarios_minimos'
    }
}


dfs = []
for uf in UFS:
    print(f'Processando {uf}')
    dfs_uf = []
    for file in model_vars:
        file_path = f'data/censo_2010/csvs/{file}_{uf}.csv'
        fields = list(model_vars[file].keys())
        fields.append('Cod_setor')
        try:
            df = pd.read_csv(file_path, delimiter=';', usecols=fields, dtype="string", encoding='UTF-8').set_index('Cod_setor')
        except:
            try:
                df = pd.read_csv(file_path, delimiter=';', usecols=fields, dtype="string", encoding='ISO-8859-1').set_index('Cod_setor')
            except:
                df = pd.read_csv(file_path, delimiter=',', usecols=fields, dtype="string", encoding='UTF-8').set_index('Cod_setor')
        df.rename(columns=model_vars[file], inplace=True)
        dfs_uf.append(df)
        # break
    df = pd.concat(dfs_uf, axis=1)
    df.loc[:, 'nome_UF'] = uf
    dfs.append(df)

df = pd.concat(dfs)

df

df_areas = pd.read_csv('resultados/area_dos_setores.zip', dtype={'CD_GEOCODI':'object'})

df_areas.set_index('CD_GEOCODI', inplace=True)

df = df.join(df_areas, how='left')

df['area_do_setor'] = df['area_do_setor'].fillna(1.)

df.shape

df.nome_rm.value_counts()

len(df.rm.value_counts().to_list())

df.dtypes

CAMPOS_MODELO = [
    'cod_grande_regiao',
    'uf',
    'nome_UF',
    'rm',
    'nome_rm',
    'meso_regiao',
    'cod_municipio',
    'nome_municipio',
    'situacao_setor',
    'tipo_do_setor',
    'numero_de_domicilios',
    'numero_moradores',
    'area_do_setor',
    'porcentagem_domicilios_sem_coleta_de_lixo',
    'porcentagem_domicilios_sem_ligacao_rede_de_agua',
    'porcentagem_domicilios_sem_banheiros',
    'porcentagem_domicilios_sem_esgoto_ou_fossa',
    # # 'porcentagem_domicilios_tipo_comodo',
    'porcentagem_domicilios_cedidos_outra_forma',
    'porcentagem_domicilios_em_outra_condicao_de_ocupacao',
    'numero_de_banheiros_por_habitante',
    'porcentagem_responsaveis_nao_alfabetizado',
    'porcentagem_responsaveis_nao_alfabetizado_com_menos_de_30_anos',
    'porcentagem_responsaveis_renda_ate_3_salarios',
    # # 'porcentagem_responsaveis_menos_8_anos_de_estudo',
    # # 'anos_medios_estudo_responsavel',
    'renda_media_responsavel',
    'numero_de_domicilios_particulares_permanentes',
    'numero_de_domicilios_improvisados',
    'numero_de_pessoas_residentes',
    'porcentagem_de_responsaveis_com_menos_30_anos',
    'numero_medio_pessoas', 
    'densidade_populacional',
    'densidade_domicilios'
]

df.loc[:, 'porcentagem_domicilios_sem_coleta_de_lixo'] = pd.to_numeric(df.loc[:, 'domicilios_com_lixo_coletado'], errors='coerce') / pd.to_numeric(df.loc[:, 'domicilios_particulares_permanentes'], errors='coerce')

df.loc[:, 'porcentagem_domicilios_sem_ligacao_rede_de_agua'] = (pd.to_numeric(df.loc[:, 'domicilios_particulares_permanentes'], errors='coerce') - pd.to_numeric(df.loc[:, 'domicilios_tipo_casa_com_abastecimento_de_agua'], errors='coerce')) / pd.to_numeric(df.loc[:, 'domicilios_particulares_permanentes'], errors='coerce')

df.loc[:, 'porcentagem_domicilios_sem_banheiros'] = pd.to_numeric(df.loc[:, 'domicilios_particulares_permanentes_sem_banheiro_ou_sanitario'], errors='coerce') / pd.to_numeric(df.loc[:, 'domicilios_particulares_permanentes'], errors='coerce')

# 'porcentagem_domicilios_sem_esgoto_ou_fossa'
df.loc[:, 'porcentagem_domicilios_sem_esgoto_ou_fossa'] = (pd.to_numeric(df.loc[:, 'domicilios_particulares_permanentes'], errors='coerce') - pd.to_numeric(df.loc[:, 'domicilios_particulares_permanentes_com_banheiro_ligado_a_esgoto'], errors='coerce') - pd.to_numeric(df.loc[:, 'domicilios_particulares_permanentes_com_banheiro_ligado_a_fossa_septica'], errors='coerce')) / pd.to_numeric(df.loc[:, 'domicilios_particulares_permanentes'], errors='coerce')


# porcentagem_domicilios_cedidos_outra_forma
df.loc[:, 'porcentagem_domicilios_cedidos_outra_forma'] = pd.to_numeric(df.loc[:, 'domicilios_particulares_permanentes_cedidos_de_outra_forma'], errors='coerce') / pd.to_numeric(df.loc[:, 'domicilios_particulares_permanentes'], errors='coerce')

# porcentagem_domicilios_em_outra_condicao_de_ocupacao
df.loc[:, 'porcentagem_domicilios_em_outra_condicao_de_ocupacao'] = pd.to_numeric(df.loc[:, 'domicilios_particulares_permanentes_em_outra_condicao_de_ocupacao'], errors='coerce') / pd.to_numeric(df.loc[:, 'domicilios_particulares_permanentes'], errors='coerce')

banheiros = [
    'domicilios_particulares_permanentes_com_1_banheiro',
    'domicilios_particulares_permanentes_com_2_banheiros',
    'domicilios_particulares_permanentes_com_3_banheiros',
    'domicilios_particulares_permanentes_com_4_banheiros',
    'domicilios_particulares_permanentes_com_5_banheiros',
    'domicilios_particulares_permanentes_com_6_banheiros',
    'domicilios_particulares_permanentes_com_7_banheiros',
    'domicilios_particulares_permanentes_com_8_banheiros',
    'domicilios_particulares_permanentes_com_mais_de_9_banheiros'
]

df.loc[:, 'banheiros'] = 0

for k,v in list(zip(range(9), banheiros)):
    df.loc[:, 'banheiros'] += pd.to_numeric(df.loc[:, v], errors='coerce') * (k+1)

# numero_de_banheiros_por_habitante
df.loc[:, 'numero_de_banheiros_por_habitante'] = df.banheiros / pd.to_numeric(df.numero_moradores, errors='coerce')

# porcentagem_responsaveis_nao_alfabetizado
responsaveis_nao_alfabetizados = (pd.to_numeric(df.loc[:, 'pessoas_responsaveis'], errors='coerce') - pd.to_numeric(df.loc[:, 'pessoas_responsaveis_alfabetizadas'], errors='coerce'))
df.loc[:, 'porcentagem_responsaveis_nao_alfabetizado'] =  responsaveis_nao_alfabetizados / pd.to_numeric(df.loc[:, 'pessoas_responsaveis'], errors='coerce')

# porcentagem_de_responsaveis_com_menos_30_anos
        
responsaveis_menos_30 = [
        'pessoas_responsaveis_com_10_anos',
        'pessoas_responsaveis_com_11_anos',
        'pessoas_responsaveis_com_12_anos',
        'pessoas_responsaveis_com_13_anos',
        'pessoas_responsaveis_com_14_anos',
        'pessoas_responsaveis_com_15_anos',
        'pessoas_responsaveis_com_16_anos',
        'pessoas_responsaveis_com_17_anos',
        'pessoas_responsaveis_com_18_anos',
        'pessoas_responsaveis_com_19_anos',
        'pessoas_responsaveis_com_20_anos',
        'pessoas_responsaveis_com_21_anos',
        'pessoas_responsaveis_com_22_anos',
        'pessoas_responsaveis_com_23_anos',
        'pessoas_responsaveis_com_24_anos',
        'pessoas_responsaveis_com_25_anos',
        'pessoas_responsaveis_com_26_anos',
        'pessoas_responsaveis_com_27_anos',
        'pessoas_responsaveis_com_28_anos',
        'pessoas_responsaveis_com_29_anos',
]

df.loc[:, 'responsaveis_menos_30'] = 0

for v in responsaveis_menos_30:
    df.loc[:, 'responsaveis_menos_30'] += pd.to_numeric(df.loc[:, v], errors='coerce')

df.loc[:, 'porcentagem_de_responsaveis_com_menos_30_anos'] = df.responsaveis_menos_30 / pd.to_numeric(df.loc[:, 'pessoas_responsaveis'], errors='coerce')

# porcentagem_responsaveis_nao_alfabetizado_com_menos_de_30_anos
alfabetizados_menos_30 = [
    'pessoas_responsaveis_alfabetizadas_de_10_a_14_anos',
    'pessoas_responsaveis_alfabetizadas_de_15_a_19_anos',
    'pessoas_responsaveis_alfabetizadas_de_20_a_24_anos',
    'pessoas_responsaveis_alfabetizadas_de_25_a_29_anos'
]

df.loc[:, 'alfabetizados_menos_30'] = 0

for v in alfabetizados_menos_30:
    df.loc[:, 'alfabetizados_menos_30'] += pd.to_numeric(df.loc[:, v], errors='coerce')

df.loc[:, 'porcentagem_responsaveis_nao_alfabetizado_com_menos_de_30_anos'] = (df.loc[:, 'responsaveis_menos_30'] - df.loc[:, 'alfabetizados_menos_30']) / df.loc[:, 'responsaveis_menos_30']

# porcentagem_responsaveis_renda_ate_3_salarios
responsaveis_renda_ate_3_salarios = [
        'responsaveis_com_renda_ate_meio_salario_minimo',
        'responsaveis_com_renda_de_meio_a_1_salario_minimo',
        'responsaveis_com_renda_de_1_a_2_salarios_minimos',
        'responsaveis_com_renda_de_2_a_3_salarios_minimos'
]

df.loc[:, 'responsaveis_renda_ate_3_salarios'] = 0

for v in responsaveis_renda_ate_3_salarios:
    df.loc[:, 'responsaveis_renda_ate_3_salarios'] += pd.to_numeric(df.loc[:, v], errors='coerce')

df.loc[:, 'porcentagem_responsaveis_renda_ate_3_salarios'] = (df.loc[:, 'responsaveis_renda_ate_3_salarios']) / pd.to_numeric(df.loc[:, 'pessoas_responsaveis'], errors='coerce')

# renda_media_responsavel
df.loc[:, ['total_rendimento_dos_domicilios', 'numero_de_domicilios']]
renda_por_domicilio = pd.to_numeric(df.loc[:, 'total_rendimento_dos_domicilios'], errors='coerce') / pd.to_numeric(df.loc[:, 'numero_de_domicilios'], errors='coerce')
df.loc[:, 'renda_media_responsavel'] = pd.to_numeric(df.loc[:, 'total_rendimento_dos_domicilios'], errors='coerce') / pd.to_numeric(df.loc[:, 'pessoas_responsaveis'], errors='coerce')


# numero_de_domicilios_particulares_permanentes
# domicilios_particulares_permanentes
df.loc[:, 'numero_de_domicilios_particulares_permanentes'] = pd.to_numeric(df.loc[:, 'domicilios_particulares_permanentes'], errors='coerce')

# numero_de_domicilios_improvisados
# total_domicilios_improvisados
df.loc[:, 'numero_de_domicilios_improvisados'] = pd.to_numeric(df.loc[:, 'total_domicilios_improvisados'], errors='coerce')

# numero_de_pessoas_residentes
# numero_moradores
df.loc[:, 'numero_de_pessoas_residentes'] = pd.to_numeric(df.loc[:, 'numero_moradores'], errors='coerce')

# numero_medio_pessoas'
df.loc[:, 'numero_medio_pessoas'] = pd.to_numeric(df.loc[:, 'numero_moradores'], errors='coerce') / pd.to_numeric(df.loc[:, 'numero_de_domicilios'], errors='coerce')

df.loc[:, 'densidade_populacional'] = pd.to_numeric(df.loc[:, 'numero_moradores'], errors='coerce') / df.loc[:, 'area_do_setor']

df.loc[:, 'densidade_domicilios'] = pd.to_numeric(df.loc[:, 'numero_de_domicilios'], errors='coerce') / df.loc[:, 'area_do_setor']

df

df.loc[:, CAMPOS_MODELO].to_csv('resultados/modelo.csv.zip', compression='zip')

df_ = pd.read_csv('resultados/modelo.csv.zip', compression='zip')

df_.columns
