==> 05-espacializando-resultados.ipynb <==
import pandas as pd
import geopandas as gpd

gdf_br = gpd.read_file('data/censo_2010/BR_Setores_2010.gpkg')

_ = gdf_br.to_crs(epsg='5880', inplace=True)

df = pd.read_csv('resultados/modelo_preditivo.csv.zip')

df.shape, gdf_br.shape

gdf_br.loc[:, 'Cod_setor'] = gdf_br.CD_GEOCODI.astype('int')

gdf_br.Cod_setor

df.Cod_setor

gdf_br_joined = gdf_br.set_index('Cod_setor').join(df.set_index('Cod_setor'), how='left', lsuffix='_df')

gdf_br_joined.loc[:, 'legenda'] = 'Setores Comuns'
gdf_br_joined.loc[gdf_br_joined.subnormal_LDA.isna(), 'legenda'] = 'Sem Informação'

subnormal = (gdf_br_joined.tipo_do_setor == 1)

gdf_br_joined.loc[subnormal, 'legenda'] = 'Subnormais'

precarios = (gdf_br_joined.subnormal_LDA == 1.0)

gdf_br_joined.loc[(precarios & ~subnormal), 'legenda'] = 'Precários'

rurais = (gdf_br_joined.TIPO == 'RURAL')

gdf_br_joined.loc[(rurais & ~precarios & ~subnormal), 'legenda'] = 'Rurais'

# gdf_br_joined.subnormal_LDA.isna().value_counts()

gdf_br_joined.legenda.value_counts()

gdf_br_joined.to_file('resultados/assentamentos_precarios_BR.gpkg', driver='GPKG')

cores = {
    "Setores Comuns": "#E9E2E7",
    "Rurais": "#C0D265",
    "Precários": "#F8BD09",
    "Subnormais": "#EA1C23",
    "Sem Informação": "#7E7C82"
}

# Apresentação do método

## Introdução
    # Objetivo rodar o modelo criado para 2010 a partir do modelo proposto pelo CEM
    # Aguardar os dados de 2022
## Método
    # Dimensões
        # as 3 dimensões que ficaram de fora
    # Modelos: LDA, RandomForest, SVM
    # agregados espacialmente: RM, UF e o restante MacroRegião
## Resultados
    # Modelo roda para todo BR
    # Performance de cada um dos modelos
    # percentual
    # Alguns exemplos
    # Resultado visual município de SP comparando com favelas, corticos, etc
## Próximos passos
    # Publicação de um Dash
        # por Município e RM
        # publicação de mapas e tabelas
