==> 06-comparando-resultados.ipynb <==
from simpledbf import Dbf5
import pandas as pd

dbf = Dbf5('data/download/SC2010_RMS_CEM.DBF')

df_CEM = dbf.to_dataframe()

# Então, pra ser bem sincero contigo, não sei ao certo qual a base de dados "final" que você está trabalhando (geralmente, não sou eu quem "disponibiliza" os dados "finais" para o site do Cem). De qualquer forma, na base "bruta" que foi preparada, a variável de interesse chamava-se "ap" e apresentava os seguintes valores:

# 1 = Aglomerados subnormais
# 2 = Assentamentos precários
# 3 = Outro tipo
# 88 = Sem domicílios particulares permanentes
# 99 = Sem informação


df_CEM.AP.value_counts()

df_CEM.TIPO.value_counts()

df = pd.read_csv('resultados/modelo_preditivo.csv.zip')

df.loc[:, 'CODSETOR'] = df.Cod_setor.apply(str)

df_joined = df_CEM.set_index('CODSETOR').join(df.set_index('CODSETOR'), how='left')

df_joined.subnormal_LDA.value_counts()

df_joined.subnormal_RandomForest.value_counts()

aps_cem = (df_joined.AP == 1) | (df_joined.AP == 2)

df_joined[aps_cem].subnormal_LDA.value_counts()

df_joined[aps_cem].subnormal_RandomForest.value_counts()

df_joined.subnormal_RandomForest.value_counts()

# df_joined.subnormal_SVC.value_counts()

df_CEM.AP.value_counts()

df_joined.loc[df_joined.subnormal_LDA == 1., 'AP'].value_counts()
