==> 07-comparacao-com-aps-pmsp.ipynb <==
import pandas as pd
import geopandas as gpd

gdf_SP = gpd.read_file('data/censo_2010/sp_setores_censitarios.zip')

gdf_SP.columns

# Codigo do municipio de São Paulo 3550308
gdf_sp = gdf_SP[gdf_SP.CD_GEOCODM == '3550308'].reset_index()

# gdf_sp.loc[:, 'Cod_setor'] = 
gdf_sp.loc[:, 'Cod_setor'] = gdf_sp.loc[:, 'CD_GEOCODI'].astype(str).astype(int)

gdf_sp.set_index('Cod_setor', inplace=True)

_ = gdf_sp.to_crs(epsg='31983', inplace=True)

df = pd.read_csv('resultados/modelo_preditivo.csv.zip').set_index('Cod_setor')

gdf_sp_joinned = gdf_sp.join(df, how='left')

gdf_sp_joinned.subnormal_LDA.value_counts()

gdf_sp_joinned.to_file('resultados/premilinar-municipio-sao-paulo.gpkg', driver='GPKG')

gdf_favelas = gpd.read_file('data/download/SIRGAS_SHP_favela.zip!SIRGAS_SHP_favela')

_ = gdf_favelas.set_crs(epsg='31983', inplace=True)

# Calcular para cada modelo:
    # as 4 matrizes de confusão
