==> 08-comparacao-PNUI-Belo-Horizonte.ipynb <==
import pandas as pd
import geopandas as gpd

gdf_MG = gpd.read_file('data/censo_2010/mg_setores_censitarios.zip')

# # Codigo do municipio de São Paulo 3106200
# gdf_bh = gdf_MG[gdf_MG.CD_GEOCODM == '3106200'].reset_index()

# gdf_sp.loc[:, 'Cod_setor'] = 
gdf_MG.loc[:, 'Cod_setor'] = gdf_MG.loc[:, 'CD_GEOCODI'].astype(str).astype(int)

gdf_MG.set_index('Cod_setor', inplace=True)

df = pd.read_csv('resultados/modelo_preditivo.csv.zip').set_index('Cod_setor')

rmbh = (df.rm == 16)

rmbh

gdf_rmbh_joinned = gdf_MG.join(df[rmbh], how='right')

gdf_rmbh_joinned.subnormal_LDA.value_counts()

gdf_rmbh_joinned.to_file('resultados/premilinar-municipio-rmbh.gpkg', driver='GPKG')
