==> 10-analises-preliminares.ipynb <==
import geopandas as gpd

gdf_br = gpd.read_file('resultados/assentamentos_precarios_BR.gpkg')

gdf_br.shape[0]

# gdf_br.tipo_do_setor.value_counts()
gdf_br = gdf_br[(gdf_br.tipo_do_setor == 0.) | (gdf_br.tipo_do_setor == 1.)]

gdf_br.legenda.value_counts()

31496 /(15815 + 31496)

gdf_br.numero_moradores.sum()

gdf_br.columns

gdf_br.loc[:, ['ID', 'legenda', 'numero_de_domicilios', 'numero_moradores']].groupby('legenda').agg({
    'ID': 'count',
    'numero_de_domicilios': 'sum',
    'numero_moradores': 'sum'
})

# gdf_br[gdf_br.rm != 0.]
gdf_br.loc[gdf_br.rm != 0., ['legenda', 'numero_de_domicilios', 'numero_moradores']].groupby('legenda').sum()

22043100.0 - 8797083.0

gdf_br.loc[:, ['rm', 'nome_rm']].groupby('rm').first()

gdf_br.loc[gdf_br.rm == 22., ['ID', 'legenda', 'numero_de_domicilios', 'numero_moradores']].groupby('legenda').agg({
    'ID': 'count',
    'numero_de_domicilios': 'sum',
    'numero_moradores': 'sum'
})

from simpledbf import Dbf5

dbf = Dbf5('data/emplasa/SC2010_RMS_CEM.dbf')

df_emplasa = dbf.to_dataframe()

# df_emplasa.columns.to_list()

precario_rf =(gdf_br.subnormal_RandomForest == 1)

subnormal = (gdf_br.tipo_do_setor == 1)

gdf_br.loc[precario_rf & ~subnormal].shape

gdf_br.loc[gdf_br.rm.isin([20., 21., 22.]) & precario_rf & ~subnormal, ['rm', 'legenda', 'numero_de_domicilios', 'numero_moradores']].groupby('rm').agg({
    'rm': 'count',
    'numero_de_domicilios': 'sum',
    'numero_moradores': 'sum'
})

df_emplasa.loc[df_emplasa.AP == 2, ['COD_RM', 'AP', 'BA_001', 'BA_002']].groupby('COD_RM').agg({
    'COD_RM': 'count',
    # 'AP': 'sum',
    'BA_001': 'sum',
    'BA_002': 'sum'
})

gdf_br.loc[gdf_br.rm.isin([20., 21., 22.]) & (gdf_br.legenda == "Precários"), ['rm', 'legenda', 'numero_de_domicilios', 'numero_moradores']].groupby('rm').agg({
    'rm': 'count',
    'numero_de_domicilios': 'sum',
    'numero_moradores': 'sum'
})

df_emplasa.loc[:, ['COD_RM', 'AP', 'BA_001', 'BA_002']].groupby('COD_RM').agg({
    'COD_RM': 'count',
    'AP': 'sum',
    'BA_001': 'sum',
    'BA_002': 'sum'
})

gdf_br.loc[gdf_br.rm.isin([20., 21., 22.]), ['rm', 'legenda', 'numero_de_domicilios', 'numero_moradores']].groupby('rm').agg({
    'rm': 'count',
    'numero_de_domicilios': 'sum',
    'numero_moradores': 'sum'
})

gdf_br.columns

gdf_br
