==> 11-analise-dados-labgeo.ipynb <==
import pandas as pd
import numpy as np
from sklearn.discriminant_analysis import LinearDiscriminantAnalysis
from sklearn.metrics import confusion_matrix

df_lab_geo = pd.read_csv('data/labGeo/dados_entrada_lab_geo-2024_01_19.csv', encoding='iso-8859-1')

df_lab_geo.columns

df_lab_gem_RMSP = df_lab_geo.loc[df_lab_geo.NOMERM == 'RM de São Paulo', ['V0003', 'V0004', 'VAR12', 
                  'HAB1', 'HAB2', 'HAB3', 'HAB4', 'HAB5', 'HAB6', 'HAB7', 
                  'CHE1', 'CHE2', 'CHE3', 'CHE4', 'CHE5', 'CHE6', 'CHE7', 
                  'HAB8', 'HAB9', 'TIPO']].fillna(0)

df_lab_gem_RMSP

df_lab_gem_RMSP.TIPO.value_counts()

X = df_lab_gem_RMSP.drop(columns=['TIPO'])
y = df_lab_gem_RMSP.TIPO

clf = LinearDiscriminantAnalysis(n_components=1)
clf.fit(X, y)

np.unique(clf.predict(X), return_counts=True)

clf.score(X, y)

cf_matrix = confusion_matrix(y, clf.predict(X))

cf_matrix

clf.predict_proba(X)

clf.predict(X)

clf.decision_function(X)
